Remove billing.find-update-invoice-accounts job
https://eaflood.atlassian.net/browse/WATER-4437

This service had a scheduled job [Licence not in charge version workflow](https://github.com/DEFRA/water-abstraction-team/blob/main/jobs/service.md#licence-not-in-charge-version-workflow). It's intent was to add a `water.charge_version_workflow` record any time it saw a new or updated licence.

Our understanding is this was intended to keep the licence from being included in a bill run until Billing & Data has had a chance to review its charge versions.

However, it appeared to have stopped working. When we dug in we realised it seemed BullMQ jobs that were scheduled using cron syntax were not always getting their repeated jobs repeated! That was resolved with [New job to add new & updated licences to workflow](https://github.com/DEFRA/water-abstraction-system/pull/903). But it caused us to check the other jobs scheduled in this way.

[Digitise To Licence Gauging Stations](https://github.com/DEFRA/water-abstraction-team/blob/main/jobs/service.md#digitise-to-licence-gauging-stations) and [Digitise to Licence Version Purpose (LVP) Sync](https://github.com/DEFRA/water-abstraction-team/blob/main/jobs/service.md#digitise-to-licence-version-purpose-lvp-sync) displayed the same sporadic behaviour when we checked the logs.

[Check for updated invoice accounts](https://github.com/DEFRA/water-abstraction-team/blob/main/jobs/service.md#check-for-updated-invoice-accounts) (otherwise known as `billing.find-update-invoice-accounts`) was never being run at all! 😱

Because it was not assigned a BullMQ queue scheduler the job was never running and had been that way for more than 2 years. On that basis after discussing it within the team it was agreed the job could be ditched altogether. 💪

This change removes the redundant job.