Remove notification jobs log noise
https://eaflood.atlassian.net/browse/WATER-4437

We recently had to investigate why a job was not running. We [eventually got to the bottom of it](https://github.com/DEFRA/water-abstraction-service/pull/2493). But one of the things that really hampered our efforts was just how much noise the notification jobs add to our logs.

For example both [Check status](https://github.com/DEFRA/water-abstraction-team/blob/main/jobs/service.md#check-status) and [Send message](https://github.com/DEFRA/water-abstraction-team/blob/main/jobs/service.md#check-status) are scheduled to run every 15 seconds. But each one logs

- when the job is being handled
- how many items are found
- that the job has been completed

That means at least 6 log entries are being made every 6 seconds. The issue is they rarely report anything! We're not a 'busy' service so 99% of the time there are no emails to send. We have to run these jobs this frequently though because of how the previous architected email sending. If we delayed the jobs folks might think there is an issue when they do eventually send a notification. 😩🤦

So, what we get in our logs is this, over and over and over and over and over .....

```
2024-04-12T17:40:00.307Z - info: Refreshing notify message events - 0 item(s) found
2024-04-12T17:40:00.308Z - info: notifications.refreshEvent: Job has completed
2024-04-12T17:40:15.724Z - info: Handling: notifications.sendMessage:repeat:fcf949af50e25a09e84caf5b78f3c43e:1712943615000
2024-04-12T17:40:15.789Z - info: Sending notify messages - 0 item(s) found
2024-04-12T17:40:15.792Z - info: notifications.sendMessage: Job has completed
2024-04-12T17:40:15.808Z - info: Handling: notifications.checkStatus:repeat:6d1e1a23af93f165a346acbe613eb94c:1712943615000
2024-04-12T17:40:15.810Z - info: Checking notify statuses - 0 item(s) found
2024-04-12T17:40:15.811Z - info: notifications.checkStatus: Job has completed
2024-04-12T17:40:30.164Z - info: Handling: notifications.sendMessage:repeat:fcf949af50e25a09e84caf5b78f3c43e:1712943630000
2024-04-12T17:40:30.181Z - info: Sending notify messages - 0 item(s) found
2024-04-12T17:40:30.183Z - info: notifications.sendMessage: Job has completed
2024-04-12T17:40:30.255Z - info: Handling: notifications.checkStatus:repeat:6d1e1a23af93f165a346acbe613eb94c:1712943630000
2024-04-12T17:40:30.258Z - info: Checking notify statuses - 0 item(s) found
2024-04-12T17:40:30.260Z - info: notifications.checkStatus: Job has completed
```

This change updates the 3 noisy jobs. Gone are the 'handling' and 'job completed' messages. We also only report items found when there are items.