Add column to `licences` table to flag for 2PT supplementary billing https://eaflood.atlassian.net/browse/WATER-4216
Part of the ticket to remove a two-part tariff licence from a bill run is to flag that licence for supplementary billing once removed.

Currently there is no column that exists in the database to flag a licence for two-part tariff supplementary billing. So in this PR a migration will be created to add it to the licence table.