Handle rtn. req. not existing when mapping rtns
https://eaflood.atlassian.net/browse/WATER-4465

We had a call that when attempting to view a licence the users were seeing an error. Initially, we thought it was the usual issue of a second record with leading or trailing whitespace. But the problem was actually missing data.

When the view licence page is loading the legacy code is fetching all return logs for a licence (though the page only displays first ten results we suspected this was the case!)

What we were not aware of is it then tries to map each return log returned to a record in `water.return_requirements`. Nothing from that record is displayed as part of the view licence page and as far as we know, no functionality uses it.

But for reasons we don't know, the records are imported and the returns mapper in this service attempts to find the matching return requirement!

Anyway, when we first joined the team we had it in mind to start improving and resolving the many issues we found in the legacy code. One of these was the use of Lodash for pretty much anything!

Its [find()](https://lodash.com/docs/4.17.15#find) was being used in the mapper to identify the matching return requirement. The code as written appears to assume there will always be a result.

So, as part of our efforts to ditch Lodash we replaced `find()` with the [standard JavaScript find()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/find).

All seemed well until the users hit on the one licence where there isn't a return requirement for every return log! The Lodash `find()` will silently eat the error. The standard library will throw if no result is found.

Rather than go back to Lodash, this change wraps the call in a try/catch. We now know it doesn't matter if the function returns undefined. Testing has confirmed the page will load. So we can catch the exception, log the fact this has happened and then let the function continue as it did before.

We're are only a couple of weeks away from replacing the legacy view licence page. So, we see this a stop-gap solution until then anyway.