Fix missing licence holder crm_v2 records
https://eaflood.atlassian.net/browse/WATER-4433
https://eaflood.atlassian.net/browse/WATER-4434

When working on the new view licence page we have been working under the premise that the licence holder for all licences is known. Unfortunately, during testing we found this is not the case. At the time of creating this fix migration there are 38 licences that do not have the correct record in `crm_v2.document_roles` to denote who the licence holder is.

When we looked at how the legacy service behaved we found that in some cases it _did_ show a licence holder! Further digging found that of the 38, 20 have the information stored in `crm.document_header.metadata`. Why these 20 never got a `crm_v2.document_roles` created we're chalking up to the brittleness of the legacy code.

This change is a migration to add in the missing records for the 20.

Whilst digging we also found there were 2 more `crm_v2.documents` (licences by another name) than exists in `water.licences`. These we found had invalid licence ref's. We speculate that the import permits the creation of the `crm_v2.documents` records but failed when attempting to insert into `water.licences`.

We've confirmed they no longer exist in the NALD data we receive so clean them out as part of this fix script.