Add point ID to rtn req points table & fix typo
https://eaflood.atlassian.net/browse/WATER-4493

Returns submitted to the service are sent back to NALD. There are then data reports that link the return submissions to their relevant points. But when we take over the returns functionality from NALD the link between a return submission (form log) and the point will be lost if we don't have some way of capturing the NALD unique point ID.

The submissions (returns.returns) link to `water.return_requirements` which in turn is linked to `water.return_requirement_points`. If that table contains the NALD point ID then they should be able to link submissions to specific points when creating reports.

This change adds a new column to the `water.return_requirement_points` table to hold the NALD point ID.

We also fix a typo in one of the other columns (`fixty_six_exception -> fifty_six_exception`) we added recently to the `water.return_requirements` table.