Add created_by to return versions table
https://eaflood.atlassian.net/browse/WATER-4522

> Part of the work to replace NALD for handling return requirements

When we take control of the setup of return requirements we'll be persisting them to the same tables that they are currently imported into.

One omission though is that there is no means to identify who created the return version. Admittedly there is nothing in NALD to indicate this hence nothing to import. But we intend to capture this information when we start persisting the records.

This adds a migration to add `created_by` to `water.return_versions`.