Update water-abstraction-helpers package https://eaflood.atlassian.net/browse/WATER-4597
Some updates to the packages in the `water-abstraction-helpers` repo were made a while ago in PR https://github.com/DEFRA/water-abstraction-helpers/pull/269

This package isn't currently set to auto-update in the `package.json` for minor/patch version changes to the helpers package. So in this PR, the helpers package will be updated and set to auto-update for minor/patch updates in the future.