Remove integrated test files (#2592)
https://eaflood.atlassian.net/browse/WATER-4523

As part of our work to replace the legacy code, we’ve tackled the complexities and failures of the old test data loader that hindered our acceptance tests. Now that all the acceptance tests have been migrated, it's time to say goodbye to the old data. This PR is removing all references to acceptance tests and their data in the service repo.