Add mod_log fields 2 return, licence & charge ver.
https://eaflood.atlassian.net/browse/WATER-4563
https://eaflood.atlassian.net/browse/WATER-4564
https://eaflood.atlassian.net/browse/WATER-4565

> Part of the work to display a licence's history to users (mod log)

NALD has a concept called 'mod log'. Essentially, when someone creates a new licence, charge or return version they are required to provide a reason and can also add a note.

Rather than this being stored against each record, the 'mod log' is stored in a single place. Users can then view a licence's 'mod logs' to get a sense of the history of the licence.

When charging switched from NALD to WRLS, we (the previous team) built the ability to record a reason and note against a new charge version, but didn't import any historic data. With return requirements also soon to switch from NALD to WRLS there is a concern that this view of changes made to a licence will be lost.

Having reviewed how the information is held in NALD we've confirmed that we can extract this information and import it as part of the licence import. The first step is ensuring we have somewhere to put it. This change adds a migration that adds a new `mod_log` field to `charge_versions`, `licence_versions`, and `return_versions`. The intent is to retain the source mod log data so we can extract details from it for historic records (imported from NALD). New charge and return versions created in WRLS already capture this data in other ways.