Add System config for Supplementary billing flags
https://eaflood.atlassian.net/browse/WATER-4622

As part of the work for two-part tariff supplementary billing, we need to add in system config for the service repo to hit a new endpoint we have created. The endpoint is to flag a licence for supplementary billing. As there are 5 different ways a licence can get flagged with 3 of them triggered from the service repo, this PR will allow us to hit the system endpoint and add the flag if necessary.