Add migration for licence versions and licence versions purposes (#2605)
* Add migration for licence versions and licence versions purposes

https://eaflood.atlassian.net/browse/WATER-4579

Alters the water.licence_versions and water.licence_version_purposes to set the date_created and date_updated to default now.

This has been done as part migrating the legacy import licence service to the water-abstraction-system. Rather than adding this logic in the system code it makes sense to add the migration and have the database handle this logic for us.

* Add migration for licence versions and licence versions purposes

https://eaflood.atlassian.net/browse/WATER-4579

Alters the water.licence_versions and water.licence_version_purposes to set the date_created and date_updated to default now.

This has been done as part migrating the legacy import licence service to the water-abstraction-system. Rather than adding this logic in the system code it makes sense to add the migration and have the database handle this logic for us.

* Add migration for licence versions and licence versions purposes

https://eaflood.atlassian.net/browse/WATER-4579

Alters the water.licence_versions and water.licence_version_purposes to set the date_created and date_updated to default now.

This has been done as part migrating the legacy import licence service to the water-abstraction-system. Rather than adding this logic in the system code it makes sense to add the migration and have the database handle this logic for us.

* Update migrations/sqls/20240814103242-alter-licence-version-purposes-created-and-updated-at-down.sql

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>

* Update migrations/sqls/20240814102304-alter-licence-version-created-and-updated-at-down.sql

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>

---------

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>