https://eaflood.atlassian.net/browse/WATER-4632
We received a report that the external app's change email journey was broken. We quickly confirmed this.

The [first part of the fix](https://github.com/DEFRA/water-abstraction-tactical-idm/pull/886) was some mis-ordered logic in the **water-abstraction-tactical-idm** project.

Having spotted that we thought we were done. But instead the feature was breaking at the point you entered the security code.

We tracked it down to an issue in this project. It seems we create an `water.event` record upon confirmation of the change being verified. Or we would if it didn't blow up.

First spot was that it was using a deprecated method for creating the event (the method and the deprecation were all created by the previous team!)

It was also using [Bookshelf](https://bookshelfjs.org) to save the record. By this point we were a little fed-up so forgive for not digging into the history. But the problem was the helper was generating the data object with an `event_id`. That event object was then getting passed to [Bookshelf's save() method](https://bookshelfjs.org/api.html#Model-instance-save). In **Bookshelf** `save()` is used for both updates and inserts. It seems **Bookshelf** determines which by checking if the thing its saving has an ID. If it does it uses update. No ID, then it uses insert.

Setting the ID in the object was making **Bookshelf** think it was doing an update. However, when the DB returned 0 for the update count **Bookshelf** was blowing up. Much like [the fix](https://github.com/DEFRA/water-abstraction-tactical-idm/pull/886), this appears to have been a broken journey from the start.

So, in this change we first switch to the non-deprecated process for creating event records. We then ensure that the helper doesn't generate an ID, which means **Bookshelf** makes an insert call instead.

All seems fine locally but we're adding a [new external user change email test](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/131) just to be sure.