Handle failed request to system supplementary (#2611)
https://eaflood.atlassian.net/browse/WATER-4622

> Part of the work to flag licences for SROC 2PT supplementary billing

We recently added support for sending a request to the [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) to let it know a new charge version had been added. The intent is for **water-abstraction-system** to determine if the licence needs to be flagged for SROC two-part tariff supplementary billing because of this.

The only problem is we ain't built the endpoint yet! 🤦

We've just realised we've broken the create charge version journey because the request is getting a failed response (404). We're not far away from adding the endpoint. But even so, we don't want a failure to flag the licence to cause the charge version record to get messed up.

This change adds a `try/catch` around the request so we can log and swallow any failures and let the legacy code continue doing its job.