Alter table default for versions `mod_log` field
https://eaflood.atlassian.net/browse/WATER-4563
https://eaflood.atlassian.net/browse/WATER-4564
https://eaflood.atlassian.net/browse/WATER-4565

> Part of the work to display a licence's history to users (mod log)

In [Add mod_log fields 2 return, licence & charge ver.](https://github.com/DEFRA/water-abstraction-service/pull/2598) we added a data migration to add a new `mod_log` field to the charge, licence and return version tables in the `water` schema. This was intended to hold details from the linked NALD mod log record, such as who, when and why the version was created.

However, we've since learnt that it is possible for a version in NALD to have multiple mod log records. Therefore, we want to change the default to avoid complexity in the code. Now we want it to default to an empty JSONB array rather than an empty object.

This change is the data migration for that.