Delete errant licence version
https://eaflood.atlassian.net/browse/WATER-4625

We had a report that when creating a new charge version for licence **AN/033/0058/018/R01** and selecting use abstraction data, the service was using an old abstraction volume.

We tracked the issue down to the fact there are 2 'current' licence versions linked to the licence (only one should ever be current). When the user selects the 'use abstraction data' option the legacy code is selecting the first 'current' licence version connected to the licence. This explains why the value is coming from an older version.

We think the version got in there from a 'blip', probably caused by the licence in NALD being duplicated and _then_ corrected with the updated data. We think the import happened in the middle of this hence the 'blip'.

Simply put, the errant licence version actually belongs to the licence that was duplicated but all the licence versions for that are correct. So, to get **AN/033/0058/018/R01** back into shape we just need to delete the errant licence version.