Remove presroc supp. billing flag migration
https://eaflood.atlassian.net/browse/WATER-4644

There has been a long-standing issue with the old scheme and the legacy code, which has resulted in large numbers of licences being flagged for presroc supplementary billing. There is no front-end way to remove billing flags, by design, to maintain an audit trail of licences that should be billed.

B&D have completed work to identify licences on a list provided by us where they have a billing flag but are billed to date. They've asked if we can remove the billing flag because there is no other way to do so.

We've done this before (see <https://github.com/DEFRA/water-abstraction-service/pull/2300> and WATER-4160) and we thought then this would be the last time we'd have to do this. Clearly we were wrong!

We suspect there is something going on either with pre-sroc billing not removing the flags, or the import is still flagging stuff it shouldn't. Either way, the sooner the legacy code is dead the better!

Till then, here is another migration. 😩