Remove presroc supp. billing flag migration (#2623)
https://eaflood.atlassian.net/browse/WATER-4644

A long-standing issue with the old scheme and the legacy code has resulted in large numbers of licences being flagged for presroc supplementary billing. By design, there is no front-end way to remove billing flags to maintain an audit trail of licences that should be billed.

B&D has completed work identifying licences that have a billing flag but are billed to date. They've asked if we can remove the billing flag because there is no other way to do so.

We've done this before (see <https://github.com/DEFRA/water-abstraction-service/pull/2300> and WATER-4160), and we thought then this would be the last time we'd have to do this. We were wrong!

We suspect something is going on, either with pre-sroc billing not removing the flags or with the import still flagging stuff it shouldn't. Either way, the sooner the legacy code is dead, the better!

Till then, here is another migration. 😩