Add new migration script
This will create the new sources table then populate it.

Then it will create the new `points` table and update the existing points tables ready for linking to it.

We don't drop any redundant columns in this migration because there will be a period where we have to wait for the [changes we've made to the **water-abstraction-import**](https://github.com/DEFRA/water-abstraction-import/pull/1028) to take affect.