Add new points and sources tables & amend existing (#2638)
https://eaflood.atlassian.net/browse/WATER-4645

> Part of the work to migrate return versions from NALD to WRLS

We recently added a new `water.licence_version_purpose_points` table to support our work migrating management of return versions to WRLS from NALD. Users need to be able to select from a licence's points when setting up a new return version and we found the source the legacy code was using (`permit.licence`) was compromised because it is not always populated during the import process.

That solved our problems when setting up a new return version. We then intended to remove _all_ references to `permit.licence` when retrieving points information, for example, when viewing the licence summary. But we became blocked because we are also extracting the source from the first point in `permit.licence`.

We didn't factor that into our new points tables (this includes `water.return_requirement_points`). We also knew NALD works with the concept of a distinct list of points, that are then linked to licence version purposes and return version requirements. We were hoping to avoid working in this way. But not only are the source details linked to the distinct points, there is other information we are missing.

So, this change covers two aspects

- add new tables to store sources and points
- update our existing tables to link to the new `water.points`

Then, we'll make changes to [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) to ensure the `water.points` table is updated during import, and the existing tables are correctly linked to it.