Data fix licence with wrong region
https://eaflood.atlassian.net/browse/WATER-4686

Back in June `WA/055/0015/022` was flagged to us as having been raised twice in NALD and therefore causing an issue with the import. The duplication was because the licence was first setup against the Midland region, and then again against Wales. Wales is the correct region, so they deleted that one in NALD when the problem was spotted.

However, because of the nightly import, WRLS's licence record was setup against the Midland region. We don't update region when importing licence information, so when the record was deleted in NALD our version stayed with the Midland region. Since then a Midland annual bill run has been created and sent.

So, to fix the licence we need to

- correct the region the `water.licences` is against
- update `external_id` in `water.licence_versions` to link to the correct NALD records
- update `external_id` in `water.licence_version_purposes` to link to the correct NALD records
- update `external_id` in `water.licence_version_purpose_conditions` to link to the correct NALD records
- update `external_id` in `water.return_versions` to link to the correct NALD records
- update `external_id` and `legacy_id` in `water.return_requirements` to link to the correct NALD records
- update `external_id` in `water.return_requirement_purposes` to link to the correct NALD records
- update `external_id` in `water.return_requirements` to link to the correct NALD records
- update `external_id` in `crm_v2.documents` to link to the second NALD licence record
- update `system_internal_id` in `crm.document_headers` to link to the second NALD licence record
- update `licence_id` in `permit.licence` to link to the second NALD licence record
- update `licence_data_value` in `permit.licence` to have the correct NALD IDs

We then need to create new charge information. This is because the header `water.charge_versions` record is also linked to a region. We can't delete or correct what is there because it is linked to a 'sent' bill. So, we duplicate the existing charge information, assign the correct region, and then mark the existing as 'superseded'.

As a data fix we need to do it as a migration. And, as most of the changes need to happen in the `water` schema we do them all here rather than split the change across the various repos.