Data fix licence with wrong region (#2639)
https://eaflood.atlassian.net/browse/WATER-4686

Back in June, `WA/055/0015/022` was flagged to us as having been raised twice in NALD and, therefore, causing an issue with the import. The duplication was because the licence was first set up against the Midland region and then again against Wales. Midlans is the correct region, so they deleted that one in NALD when the problem was spotted.

However, because of the nightly import, WRLS's licence record was set up against the Midland region. We don't update the region when importing licence information, so when the record was deleted in NALD, our version stayed with the Midland region. Since then, a Midland annual bill run has been created and sent.

So, to fix the licence, we need to

- correct the region the `water.licences` is against
- delete any `water.licence_versions` linked to the deleted NALD licence
- delete any `water.licence_version_purposes` linked to the deleted NALD licence
- delete any `water.licence_version_purpose_conditions` linked to the deleted NALD licence

We also need to create new charge information. The header `water.charge_versions` record is also linked to a region. We can't delete or correct what is there because it is linked to a 'sent' bill. So, we duplicate the existing charge information, assign the correct region, and then mark the existing as 'superseded'.

> Note: We checked return versions (there are none), `crm.document_headers`, `crm_v2.documents` and `permit.licence` records and all are linked to and being updated from the correct NALD licence record