Revert "Data fix licence with wrong region" (#2640)
Revert "Data fix licence with wrong region (#2639)"

This reverts commit 21d4ee8f753bfa3583f224b9dcbfdb5c4146922e.

Reverts DEFRA/water-abstraction-service#2639

Our fantastic QA team have raised the question of what happens when the B&D team create a new supplementary bill for Wales that includes this licence.

Because the old bill will be in a different region it will not be seen and therefore applied as a credit to the new bill that is generated.

This means the customer will be invoiced for the full amount, when they should only be invoiced/credited for the difference the change has resulted in.

We do not believe what we are doing with our data fix will be any different once this has been resolved. But we don't want to run the risk of it being applied until this is cleared up.