Add new methods (of abstraction) table for points
https://eaflood.atlassian.net/browse/WATER-4645

> Part of the work to migrate return versions from NALD to WRLS

We recently [Added new points and sources tables & amend existing](https://github.com/DEFRA/water-abstraction-service/pull/2638) to the service. Not only does it support our work on return versions, it means we can move away from abstracting point information from the licence JSON blob in `permit.licence`.

We're looking to also migrate the pages linked to from the view licence summary, which are currently in the legacy [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui). One of them is information on points. We thought we'd captured everything 'point' related in the change but have since spotted 'Means of abstraction'.

This is linked to a point in NALD and gives how water will be abstracted there, for example, "Surface Mounted Pump (Fixed)". We need to capture this information as we import the point information from NALD. So, the first step is to add the table and update `water.points` so that it can be linked to a method.

> We have to name our table and entities in code and the DB. `means` doesn't really mean anything, and `means_of_abstraction` causes problems because it is not plural. Hence `methods`!